requirements.txt
flask
datetime
pyserial
